:construction: WIP creating higher order component for tooltips
